requirements.txt
python-binance==0.7.10

websocket~=0.2.1